% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{coef.hglm}
\alias{coef.hglm}
\title{Title coef method for hglm object}
\usage{
\method{coef}{hglm}(hglm_out, ...)
}
\arguments{
\item{hglm_out}{the hglm object}

\item{...}{further arguments passed to or from other methods}
}
\value{
the estimated coefficients of the model
}
\description{
Title coef method for hglm object
}
