% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.R
\name{coef.hiperres}
\alias{coef.hiperres}
\title{Title coef method for hiperres object}
\usage{
\method{coef}{hiperres}(result, ...)
}
\arguments{
\item{result}{the hiperres object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Title coef method for hiperres object
}
