% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{print.hglm}
\alias{print.hglm}
\title{Title print method for hglm object}
\usage{
\method{print}{hglm}(hglm_out, ...)
}
\arguments{
\item{hglm_out}{the hglm object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Title print method for hglm object
}
