% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_analysis.R
\name{print.hiperres}
\alias{print.hiperres}
\title{Title print method for hiperres object}
\usage{
\method{print}{hiperres}(result, ...)
}
\arguments{
\item{result}{the hiperres object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Title print method for hiperres object
}
