% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_output.R
\name{vcov.hglm}
\alias{vcov.hglm}
\title{Title vcov method for hglm object}
\usage{
\method{vcov}{hglm}(hglm_out, ...)
}
\arguments{
\item{hglm_out}{the hglm object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Title vcov method for hglm object
}
