% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiper_glm.R
\name{hiper_glm}
\alias{hiper_glm}
\title{Title the core fitting function for hiperglm package}
\usage{
hiper_glm(design, outcome, model, option)
}
\arguments{
\item{design}{design matrix}

\item{outcome}{outcome vector}

\item{model}{model to fit}
}
\value{
an object of hglm
}
\description{
Title the core fitting function for hiperglm package
}
